Split main.css into naut base and our custom CSS
This commit separates the portion of Naut 3 that we employ as our base
CSS and the custom CSS we add to make the subreddit. They are now
located in:
- naut_base.scss
- custom.scss

Future additions to the subreddit stylesheet should be made in
custom.scss.

Also included is a Makefile that will use sass to concatenate the files
and minify them into main.css, which is ready for use on the subreddit.